split stats controller
