+++ app/controllers/stats/public_controller.rb
@@ -1,0 +1,7 @@
+module Stats
+  class PublicController < BaseController
+    def index
+      @stats = Stats::Public::All.new(stats_params)
+    end
+  end
+end
