+++ app/views/pages/_footer.haml
@@ -24,7 +24,7 @@
       %li= link_to t('cgu'), cgu_path
       %li= link_to t('about.mentions_d_information.title'), mentions_d_information_path
       %li= link_to t('about.mentions_legales.title'), mentions_legales_path
-      %li= link_to t('usage_stats'), stats_path
+      %li= link_to t('usage_stats'), public_index_path
       - if current_user&.is_admin?
-        %li= link_to t('usage_team_stats'), team_stats_path
+        %li= link_to t('usage_team_stats'), team_index_path
       %li= link_to t('.dinum'), 'https://www.numerique.gouv.fr/dinum/'
