+++ app/views/stats/_stats_charts.haml
@@ -1,4 +1,4 @@
 - cache("stats_charts-#{stats.params}", expires_in: 1.hour) do
   #stats-charts
     - @charts_names.each do |name|
-      = render 'stats_chart', stats: stats, name: name
+      = render 'stats/stats_chart', stats: stats, name: name
