+++ app/views/stats/public/index.haml
@@ -11,8 +11,8 @@
 %section.section.section-grey#section-stats
   .container.container-full
     - cache("stats_charts-#{@stats.params}", expires_in: 1.hour) do
-      = render 'stats_params', stats: @stats, route: stats_path
+      = render 'stats/stats_params', stats: @stats, route: public_index_path
       #stats-charts
         - names = [:solicitations, :solicitations_diagnoses, :exchange_with_expert, :taking_care, :themes, :companies_by_employees, :companies_by_naf_code]
         - names.each do |name|
-          = render 'stats_chart', stats: @stats, name: name
+          = render 'stats/stats_chart', stats: @stats, name: name
