+++ app/views/stats/team/_tabs.html.haml
@@ -1,11 +1,11 @@
 .tiles
   .grid
-    = active_link_to quality_stats_path, class: 'tile' do
+    = active_link_to quality_team_index_path, class: 'tile' do
       %i.fas.fa-award.tile__icon
       %h2= t('.quality')
-    = active_link_to matches_stats_path, class: 'tile' do
+    = active_link_to matches_team_index_path, class: 'tile' do
       %i.fas.fa-handshake.tile__icon
       %h2= t('.matches')
-    = active_link_to deployment_stats_path, class: 'tile' do
+    = active_link_to deployment_team_index_path, class: 'tile' do
       %i.fas.fa-rocket.tile__icon
       %h2= t('.deployment')
