+++ config/locales/views.fr.yml
@@ -586,6 +586,9 @@ fr:
   stats:
     archived: Besoins abandonnés
     other_status: Autres status
+    public:
+      index:
+        title: Statistiques d’utilisation
     series:
       advisors:
         title: Conseillers publics
@@ -634,8 +637,6 @@ fr:
       themes:
         subtitle: Une thématique est un regroupement de sujets d’entreprises proches. <br>Le pic que l’on observe au mois de mars 2020 dans la catégorie « Autres typologies de besoins » correspond aux demandes exprimant des difficultés liées au covid-19
         title: Thématiques choisies par les entreprises
-    show:
-      title: Statistiques d’utilisation
     stats_params:
       end_date: Date de fin
       region: Région
@@ -678,14 +679,15 @@ fr:
     status_done_no_help: Clôture sans accompagnement
     status_done_not_reachable: Entreprises injoignables
     status_not_for_me: Besoins refusés
-    tabs:
-      deployment: Déploiement
-      matches: Mises en relation
-      quality: Suivi qualité
     taken_care_after: Plus de 5 jours
     taken_care_before: Moins de 5 jours
     team:
-      title: Statistiques équipe
+      index:
+        title: Statistiques équipe
+      tabs:
+        deployment: Déploiement
+        matches: Mises en relation
+        quality: Suivi qualité
     with_diagnoses: Demandes transmises
     without_diagnoses: Demandes non transmises
   usage_team_stats: Statistiques équipe
